Rename to TabularData, pull out column into a separate file
