+++ app/assets/stylesheets/webapp/tabular_data.scss
@@ -1,4 +1,4 @@
-.listing {
+.tabular-data {
   width: 100%;
   background: #fff;
   margin: 20px auto;
