+++ app/views/proposals/archive.html.erb
@@ -6,7 +6,7 @@
   </div>
   <div class="row">
     <div class="col-sm-12">
-      <%= render partial: "shared/listing", locals: {listing: @listing} %>
+      <%= render partial: "shared/table", locals: {container: @proposal_data} %>
     </div>
   </div>
 </div>
