+++ config/tables/proposals/default.yml
