+++ config/tables/proposals/ncr.yml
